✨(frontend) store import task id per mailbox (#411)
Currently when we store a task import id globally. So once the user has
launched one import he is not able to start a new import from another
mailbox. Furthermore, currently the storage never expires that is weird.
Now after 48hours, if the task is still pending we consider it as stale.